refactor: 최종우승자 선정 후 화면 표시
decideWinners 함수 GameController 클래스 내부로 이동
showWinners 함수 View 클래스 내부로 이
